Rename osu.Game/Screens/Multiplayer to Multi.
