+++ osu.Game/Screens/Multi/DrawableRoom.cs
@@ -18,7 +18,7 @@ using osu.Game.Graphics.Sprites;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Users;
 
-namespace osu.Game.Screens.Multiplayer
+namespace osu.Game.Screens.Multi
 {
     public class DrawableRoom : OsuClickableContainer
     {
