+++ osu.Game/Screens/Multi/RoomInspector.cs
@@ -21,7 +21,7 @@ using osu.Game.Graphics.Sprites;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Users;
 
-namespace osu.Game.Screens.Multiplayer
+namespace osu.Game.Screens.Multi
 {
     public class RoomInspector : Container
     {
